modify api/ctqmc_api.f90, fix bug in get_grnf()

+++ src/ctqmc/api/ctqmc_api.f90
@@ -736,10 +736,10 @@
 
 ! external arguments
 ! size of grnf
-     integer, intent(in)      :: size_t
+     integer, intent(in)     :: size_t
 
 ! impurity green's function
-     complex(dp), intent(out) :: grnf_t(size_t)
+     complex(8), intent(out) :: grnf_t(size_t)
 
 ! declare f2py directives
 !F2PY intent(in) size_t
